completely replacing old style thing for statblocks
